Use uintptr_t type when casting pointers to integers to avoid -Wpointer-to-int-cast warnings.

+++ src/server/lua/llimits.h
@@ -10,6 +10,7 @@
 
 #include <limits.h>
 #include <stddef.h>
+#include <stdint.h>
 
 
 
@@ -62,7 +63,7 @@ typedef unsigned long lint32;  /* unsigned int with at least 32 bits */
 ** conversion of pointer to int (for hashing only)
 ** (the shift removes bits that are usually 0 because of alignment)
 */
-#define IntPoint(p)  (((unsigned long)(p)) >> 3)
+#define IntPoint(p)  (((uintptr_t)(p)) >> 3)
